+++ conf/url_redirects.py
@@ -8,6 +8,12 @@ from core.views import (
 
 redirects = [
     url(
+        r'^new-zealand-event-calendar/$',
+        QuerystringRedirectView.as_view(
+            url='https://www.events.great.gov.uk/ehome/trade-events-calendar/all-events?keyword=zealand'),
+        name='new-zealand-redirect'
+    ),
+    url(
         r'^story/online-marketplaces-propel-freestyle-xtreme-sales/$',
         QuerystringRedirectView.as_view(url=reverse_lazy('landing-page')),
         name='casestudy-online-marketplaces'
@@ -239,10 +245,6 @@ redirects = [
         QuerystringRedirectView.as_view(
             url='https://selling-online-overseas.export.great.gov.uk',
         ),
-    ),
-    url(
-        r'^trade/$',
-        QuerystringRedirectView.as_view(url='https://trade.great.gov.uk'),
     ),
     url(
         r'^uk/privacy-policy/$',