+++ conf/urls.py
@@ -2,6 +2,7 @@ from directory_constants import slugs
 
 import directory_components.views
 from directory_components.decorators import skip_ga360
+from directory_constants.urls import build_great_international_url
 import directory_healthcheck.views
 
 from django.conf import settings
@@ -143,10 +144,7 @@ urlpatterns = [
     ),
     url(
         r"^export-opportunities/$",
-        RedirectView.as_view(
-            url=settings.SERVICES_EXOPPS_ACTUAL,
-            permanent=False
-        ),
+        RedirectView.as_view(url=settings.SERVICES_EXOPPS_ACTUAL),
         name='export-opportunities'
     ),
     url(
@@ -636,6 +634,35 @@ ukef_urls = [
     ),
 ]
 
+international_redirects_urls = [
+    url(
+        r'^trade/$',
+        RedirectView.as_view(url=build_great_international_url('trade/'), query_string=True),
+        name='international-trade-home'
+    ),
+    url(
+        r'^trade/(?P<path>[\w\-/]*)/$',
+        RedirectView.as_view(url=build_great_international_url('trade/incoming/%(path)s'), query_string=True),
+        name='international-trade'
+    ),
+    url(
+        r'^investment-support-directory/$',
+        RedirectView.as_view(
+            url=build_great_international_url('investment-support-directory/'),
+            query_string=True
+        ),
+        name='international-investment-support-directory-home'
+    ),
+    url(
+        r'^investment-support-directory/(?P<path>[\w\-/]*)/$',
+        RedirectView.as_view(
+            url=build_great_international_url('investment-support-directory/%(path)s'),
+            query_string=True
+        ),
+        name='international-investment-support-directory'
+    ),
+]
+
 urlpatterns += legacy_urls
 urlpatterns += deprecated_urls
 urlpatterns += euexit_urls
@@ -647,3 +674,4 @@ urlpatterns += marketaccess_urls
 urlpatterns += community_urls
 urlpatterns += ukef_urls
 urlpatterns += marketing_urls
+urlpatterns += international_redirects_urls